Truncate log file non readable end
It can be the case that on recovery log entry file reader will be not able
to read log file till the end because LogEntry was not written completely,
for example. In those cases, we need to trim those bytes because it can happen
that if we will start appending after that to a log file and will try to
re-read this half stored record again - we will be able to read it (since
bytes will be there), but the content of that record will be absolute gibberish.
To prevent that from happening we need to make sure that our log files are
always readable from a to z and trim anything that we can't read.