Fix naming bug in optional match between aggregations
This is a patch. We want to improve all of scoping and binding, which
should improve things here a lot. I'm not certain that this bug could not
resurface in a more complex query, but it's essentially a symptom of
tech debts in this area, and not something that is easy to fix in an
isolated fashion.