Fix more acceptance tests in slotted runtime
Most failures cause by `ExecutionContext.copyWith` which creates a new
`ExecutionContext` with the slot configuration of the old one. This wrong
as new `ExecutionContext`s have to use the slot configuration of the
current operator, because these could be different. Creating
`ExecutionContext`s with the wrong slot configuration typically gives
bug with the error msg "Ouch, no suitable slot for key ...".