Porting of space filling curves from "Neo4j Spatial"
This is a separate module so that it can be used for future versions of the Neo4j Spatial library from which it was derived, so we don't maintain two copies.
The included Envelope and EnvelopTest are trimmed down from the Neo4j Spatial versions, so we might want to refactor that later, either include all, or have Neo4j Spatial extend this one.