Backup should clear the store id files.
Taking an online backup effectively looks like a system crash, followed by a recovery.
In such cases, the id files are not trustworthy anyway.

More importantly, however, a backup may be used to seed a cluster, and in such cases the id files must not contain any free ids that can be reused.
The reason is that id-reuse is an uncoordinated operation in clustering, and therefor no two cluster members can be allowed to have overlapping sets of free ids.
Overlapping free id sets in a cluster can, with unfortunate timings, potentially lead to inconsistencies.