Unflake CertificatesIT
Test lives in the 'server' module and is executed together with
integration tests for REST, like `AutoIndexIT`. Later often use a
shared server which is only shutdown in a JVM shutdown hook. This
made `CertificatesIT` flaky because it tried to start it's own
bolt-enabled database on an occupied default bolt port.

This commit fixes that problem by making test use different listen
address for bolt server. It is basically a backport of
dafeb4c895cebd9e99210a935ea1971126006300.