Extracted factory for bolt protocol handlers
Previously mapping from protocol version to a function that created
protocol handler was kept in a map. This map was also passed around
as-is. It was not pretty because map type was rather large and it
was mutable at any point.

This commit extracts a dedicated factory that is able to create
protocol handlers for the given version and network channel. Code
looks cleaner with this factory.