Use array when unpacking list of known size
Bolt server can unpack lists of unknown size. It stores all unpacked
values in an `ArrayList` which automatically grows as elements are
added. Then list is converted to an array. Same thing was done for
lists of known size.

This commit removes intermediate `ArrayList` so that only a single
array is allocated.