Fixes a rare race in EncodingIdMapper regarding duplicate ids
This was a problem when detecting duplicates which would later trigger a failure
when sorting the duplicate values. There were multiple detector workers which got
its own slice of the data, the trackerCache more specifically.
Detectors marks data items as collisions and the data structure is shared
between the detectors and intentionally unsynchronized. This allows a benign race
of slightly over-counting the number of collisions when doing the detection.
The rest of the duplication logic, including sorting of the duplicate values, would
use that count as a de-facto count to determine ranges and high id in the
duplicate data array. This would sometimes cause the sort algorithm to notice
unexpected ids in the high end of the spectrum which looked uninitialized
and so fail. Now instead the initial collision count is used as a pessimistic
count to instantiate the data structures capable of holding the duplicates,
but the actual collision count is used for the rest of the algorithm.
The actual count can be had by piggy-backing on the gathering of duplicate
values, so no overhead.

ImportToolTest and EncodingIdMapperTest would see this problem now and then