Removes some unnecessary indirections from embedded impl
NodeProxy/RelationshipProxy needs some sort of SPI, acting as their relationship
to the database instance. There were a couple of unnecessary indirections and
round trips for the execution paths right below them, which were apparently
not really required and only made code more complicated and weird.
This is a cleanup and simplification of that part of the code.