Fix a few cases where a static initialiser references a subclass.
These situations can lead to classloader deadlocks, if the super-class and the sub-class ends up being loaded concurrently in two different threads.
Not all cases are fixed: some a locked in place by being public API, and another was left be due to unpredictable consequences for forward merging.

(This reverts a revert of the original commit, which had to be reverted
due to a mistake.)