Use tmp byte buffer on insert and remove slot
Previously slot was moved one slot at the time using cursor.copyTo.
This turned out to be very expensive especially when a large number
items needed to be moved, when keys are small.

Instead we now use a byte array that we temporarily write to
before writing it down again. This gives us the possibility to
operate on bigger chunks.

Benchmarking will decide how big this byte array should be.