Add monitoring of "page cache usage" that will help with tuning of the page cache size.
This add the metric `neo4j.page_cache.usage_ratio` that returns a number
between `0.0` and `1.0`. For example, a value of `0.5` mean that half of
the pages are used and the rest are available for data without evictions.

It's also exposed the same metric through the JMX bean "Page cache" with
the method `getUsageRatio()`.