Fixes a pageSize remap issue in GBPTree
Which was exposed by a test in LabelScanStoreTest that randomly scrambled the header
of a GBPTree file. It so happened that now and then it would provide a negative
page size, the existing sanity checking of page size didn't check this and so would
move on to close the pagedFile and then fail when trying to map the file with the
negative page size, which would then result in the fail-safe closing in GBPTree
constructor to itself fail due to the pagedFile already being closed and so
would fail the test in an unexpected way.