Basic implementation of native schema string indexing layer
Much of the changes in here are about making test classes and utilities
generic to value type as well as copy type-specific indexing code from
number schema index classes.

Here's also a basic, non-optimized implementation of the string schema
index accessor, string layout and friends which focuses on getting the
functionality in place first and foremost.