Big refactor of relationship selection helpers
Depending on the usage pattern, it is sometimes convenient to use
an iterator returning relationship object of some type, and sometimes
better to use a cursor approach where we do not need to create extra
objects. To achieve this but only have one implementation of the
selection logic itself, we ended up with this rather complex design,
and the bastard method fetchNext with both returns a boolean, and
does a callback to allow object creation.