Delete flaky test with limited value
This tests a negative requirement and does so in a clustered setting.

After discussions we have reached the conclusion that it is better
to delete it rather than try to fix it. There is nothing cluster-
specific about this functionality and it is enough to test the
positive requirements in a single instance system.