Delete CodeGeneratorTest
This was the first ever test we wrote for the compiled runtime, and
it mocks the entire database. It might have been a good idea at the time
but now it has turned into a maintenance nightmare and it doesn't test
stuff that are not tested in numerous other places on _real_ graphs.