Fixes as issue with applying index updates to unique indexes
Given a uniqueness contraint on Foo:bar. If a transaction make changes
such that in the end, the constraint is not violated but some intermediate
individual step violates the constraint the index would still throw an
exception.

The reason this happens is that native index validates the uniqueness
constraint on every insert and can not defer this validation to later,
unlike Lucene.

We avoid this problem by first applying all removals and then adds. Also
changes are split up into a remove and an add and applied accordingly.