SharedDiscoveryService works with refuseToBeLeader
Topologies are bootstrapped by the first server that does not refuse to be leader. The bootstrapping member must become leader to upload the snapshot to the other members.

This logic also exists in HazelcastClusterTopology, but is coupled with the underlying data structures in a way that is hard to extract.