Removes special handing of 2.3 test case that does not seem flaky
This test was supposed to be flaky, but I can never get it to pass.
Since we are not fixing the bug itself, this simply keeps that
test on the blacklist and removes the special flakiness handling.

It also prevents NullPointerExceptions when catching an Exception
that has a null error message.