Allow consistency checker to warn about indexes that are dirty and can return wrong results
Dirty native index is an index that will require recovery on next non
read only startup and potentially any data that is in the index
on a moment when index is dirty are not fully trustable.

This commit allow CC to notify about those indexes during CC check and
report those.