Always attach CORS headers in REST
Previously CORS headers were only added to requests that passed auth.
This made JS requests from browser unable to receive the actual error
because CORS is not allowed with missing headers.

This commit moves the relevant filter to the top of the filter chain so
that every response has CORS headers attached.