Close procedure streams on failures
Make sure we call any close handlers attached to procedure
streams on exceptions. Exceptions that occur during close are
collected as ResourceCloseFailureException and attached as suppressed.

Introduces a ResourceManager/ResourceTracker on KernelStatement to
keep track of active resources so that any non-exhausted procedure
streams are always closed if the transaction is closed during query
execution without closing the query result first.

The ResourceManager implementation is separated into a
class CloseableResourceManager so we can easily move it if we
will not keep KernelStatement in the future.

Also improves the error message.