Create a specialised time based task scheduler.
Instead of using the ScheduledExecutorService, this new scheduler will ensure that tasks are both started in their respective target
thread pools, run with their given ThreadGroup, and do not overlap in time in the sense that one run of a recurring task will not
start until a prior run has finished. Overdue and delayed tasks are handled as well.