Better handling of degree in compiled runtime
There is no point in calculate the degree for non-dense nodes since then
that amounts to looping through the relationship chains and then you might
as well start expanding right away. So only compute degrees if both nodes
are dense, if one of them are dense start from the non-dense one or otherwise
just stick to what you've got.