Lazy evaluation of paths
There has been reported regressions because of the introduction of values
when dealing with paths. This is caused by that we eagerly fetch all
nodes and relationships of the path in order to construct the `PathValue`
no matter if we use it or not. Fix this by introduce a wrapping - just
 as we do for nodes and relationships - for paths. We will probably want
 to fix this nicer in later versions but this felt like the most non-evasive
 fix to the regression.