Introduce an UNLABELLED label lock.
This lock is then used to coordinate schema modifications to e.g. fulltext indexes that cover the _arbitrary_ set of labels.
Without this lock, it is possible to concurrently drop an all-labels fulltext index while other transactions are trying to write to this index.
We _may_ be able to get rid of this lock again, if we serialise all fulltext updates to a dedicated writer thread, but that goes against the
grain of the current schema index architecture.