Create constraint in a single transaction
Before this PR for cases when constraint was relying on index for some
validation purposes - this index was created in a separate transaction.
And since that single logical operation of constraint creation was
split into two transactions it was possible to end up in a state
when the index was created, but next transaction that will actually create
the constraint definition will never come.
This PR changes that behavior by creating constraint using single
transaction. By doing that we will complete that situation all together.