Make page cache warmer file handling more sympathetic to how backup and store copy works.
* The page cache warmer now stores its profile data in GZIP format,
which includes checksums that maintain the validity of the files.
* This way, we no longer need to do our replacing file move dance in
order to attempt to update the profiles pseudo-atomically.
* Instead, the files are overwritten in-place, and if the checksum is
wrong, we just discard the profile.
* Removing the file move dance means that backup will now no longer see
that a file it wants to copy, has been removed while it wasn't looking.
* Previously, when this happened, it would cause backup or store copy to
fail, but now it never happens so we're good.
* This also solves a problem where backup could observe a file to have
contents, which gets truncated to zero before or while backup is
streaming those contents. This could also cause backup to fail with very
strange errors, but now the files are never truncated to zero, so this
problem is never encountered.