Fixes a native uniqueness index creation issue in clustering
The observed issue is that of clustering, where activation of
uniqueness constraint could race with concurrent uniqueness-violating
transactions on other cluster members without the constraint creator
noticing. The result would be a "successfully" created constraint
containing violations to the constraint.

A couple of changes were made to native indexing to fix this
problem w/o resorting to full scan of the index to verify uniqueness:

 - Unique and non-unique layout now both share the compare method
   from the unique layout. This is so that (gb+)tree uniqueness, i.e.
   (value) or (value,entityId) can be controlled per interaction
   with the tree.
 - Index population uses (value) tree uniqueness for adding data
   from store scan and (value,entityId) w/ deferred conflict checking
   for applying updates during population.
 - TentativeIndexProxy, i.e. limbo between populating and online,
   also uses (value,entityId) w/ deferred conflict checking.
 - Online index always uses (value,entityId) uniqueness,
   even for unique indexes. Kernel is required to do
   uniqueness checking and concurrency control outside the index
   anyway so additionally checking uniqueness inside the index
   introduces problems for no gain.
 - Index seekers always use (value,entity) tree uniqueness,
   i.e. they read what's there, plain and simple.

The reason why most need to use (value,entityId) tree uniqueness
is that batches of updates may, depending on ordering, contain
intermediary conflicts that should not really be considered
conflicts until the whole batch have been applied. Also to consider
here is that batches may come in for any number of transactions
batched into one (clustering catch-up scenario).

Index population can use (value) tree uniqueness since it's only
adding entries and fail-fast is a desirable feature.

There was a change to native index populator to allow for deferred
constraint checking too, which is that the gb+tree writer isn't
open all the time, but gets opened and then closed per "work sync".