Default port fix for backup
CC and HA use different protocols on different ports for performing
backups
Because of this the default port is different for the service that is
being backed up from
To establish which default is correct it needs to evaluate just before
connecting
The reason it is failing is because that code looks at the value of the
parameter and sees that it isn't empty. The default value of the
argument is "localhost:6362" which looks like the user had typed it in.