Use LocalSlotReferences to close all hanging transaction on pool close.
Before we were using weak references and reference queue when pool was
closed to cleanup all of the objects that were allocated but not yes disposed.
From time to time that was not possible since we
were able to observe state where WeakReference was already cleaned
but not yet visible in reference queue. That would cause some of the
objects (in our case transactions) to not release their resources properly.
For example batch of allocated ids will never be returned and will be lost
until id generator file rebuild.
This PR changes that behaviour and use map of slotReferences to reliably
close all entries that available in the pool.
PR also makes pools AutoClosable.