Introduce increased generated testing of cypher using index seeks
This commit extends SemanticIndexAcceptanceTest to test Point and
Date in addition to the previous Long/Double/String. In doing this,
several minor problems were found for both code paths and corrected.

Note that this commit further extends the shaky index query code around
TransactionBoundQueryContext, which extends all the way into
PropertyValueComparison and other places. This was done to get something
which passes tests, but this area is in dire need of a rewrite. The
other temporal types should not be stacked on top of that mess.