Don't log tx terminated stack on client disconnect
When a client disconnects mid-operation, the bolt server
stack will call "halt" on the state machine. That will, in turn,
terminate any running transaction, and the error from that termination
will propagate up the stack in the worker thread, which will then
try to write the error back to the client.

TL;DR: Any time a connection is closed mid-operation, a massive
stack trace fills up the debug log.

These become reasonably common in production logs, for instance
if users of the Neo4j Browser just close the window in the middle of a query.

The problem isn't *harmless* - it could be indicative of a misconfigured
application that is exiting before a query completes, which might mean
it's losing writes.

Hence, don't get rid of the log message entirely, but replace the big
stack dump with a more helpful `warn` message explaining how to handle.