Don't write or read backups through page cache unless we have to.
Doing backup through the page cache was a necessary change to support
block devices, however, this introduced an overhead in cases where it
wasn't necessary. This commit will revert the behaviour to the old way
of doing thing, unless the underlying filesystem is not the default
one, i.e. block device file system.