Redesign StoreFileStreams
We want to avoid using append when writing to file. Since files can
now retry, we would get a situation where files would append from
the beginning to half finished files.

The StoreFileStream would previously use a kind of mixed design where
page files would be stored in a map while normal files would just open
new output stream and append.

This introduces a StoreFileStreamProvider for the catchup client. Each
StoreFileStream need to be closed and append has been disabled so that
we always start over when retrying to download a file.