Support uniqueness constraints for spatial
This is primarily a set of tests, plus one small bug-fix because it
turns out the empty native indexes initialize the cursor correctly
so they will check tx-state, while the non-existent spatial indexes
did not (because they were non-existent). We now initialize the cursor
at the point the reader is discovered to be non-existent.