Do not let index seeks throw on impossible predicates
If a predicate is statically guaranteed to be impossible, we should
still treat this as a user error and give a nice error message. This
should not be dependent on the existence or not of an index, and should
be done during linting as soon as type inference is complete.