GBPTree.readHeader don't need layout
Layout was just needed to read header as an extra safety net.
We verified that the index that you read the header from was actually
the index that you thought it was, by verifying the layout. This created
more problems than it solved and we can drive without this seat belt
holding us back.

Also aligning exception type on missing meta page and missing state page.
Use MetadataMismatchException in both cases.