Build only one 3D Hilbert space filling curve
The tree loops back on itself so there is only one instance of any
curve configuration. This means that once the tree of all sub-curves
is built, this structure supports traversal to infinite depth. Therefore
this one static structure can support all 3D curve queries for all
indexes of this envelope and dimensionality. Building the whole tree
statically also solves a concurrency problem we had with concurrent queries
racing for the lazy creation of the tree we were previously doing.