Unify behaviour of ArithmeticException across runtimes
- Math functions in the values module does not rethrow arithmetic
Java exceptions as Neo4j exceptions.
This lead to different behaviour across the runtimes, where
only the compiled runtime was doing this.
For now let us not alter the message (e.g. "long overflow") coming out
of Java, but just make sure we wrap it in a public Neo4j exception.

- Floating point division should not throw "/ by zero".
Division by zero has well-defined representations in the IEEE-754
floating point standard, and it would be rude to abort a query
unless the user has opted in to trapping floating point exceptional
conditions. Currently the JVM does not support such trapping.