Use epochSecond of LocalDateTime for ZonedDateTime in Bolt
So that values of `DateTime` type (represented with `ZonedDateTime`)
are not converted to UTC. Previously Bolt server converted
`ZonedDateTime` to epochSecond in UTC before sending it and expected to
receive value in UTC as well. This forced clients to perform a
conversion from UTC and to UTC. Such conversion can be especially hard
with named timezones. Now clients will receive unadjusted value which
can be directly converted to `LocalDateTime` and used together with
the timezone info.