Circumvent race in TransactionHandleRegistry.
There was a race between the transaction being rollbacked and the rollback http request. In terminate method, the async "handle.terminate()"" could run before the acquire, thus removing the handler from the map before it could be retrieved again. This was very evident on windows, most likely due to a different scheduling.