Make sure that handlers share state on test catchup server
This caused wrong decoders to be used in the test server. Eventually requests would
time out which meant that the channel would be replaced. Hence tests
would eventually be successful but much slower.

This refactors the catchup server handler by adding the protocol in the methods
instead of the constructor. For more easy to use.