Optimize slotted runtime result production
Make ProduceResultSlottedPipe return an iterator that can materialize
the results directly into an array of columns that can be consumed by
Bolt directly without going through a generic Map interface.
This lets us avoid unnecessary copies and hash map creations.

- Introduce an ArrayResultExecutionContext that implements
QueryResult.Record, and stores the columns in an ordered array
for fast direct consumption, with support for the old Map interface
as a secondary concern.

- Use an instance cache for ArrayResultExecutionContext to reduce the
overhead of array creation and initialization.