Fix race in netty protocol install
The message gate and the error handlers are now left in place, so that
a message being written while the pipeline is being modified cannot
race past the gate and reach nettys sentinel handler. The exact
circumstances under which this happens are not really understood
though, because the change of the pipeline was happening from inside
the event loop of the channel, which should make it safe.

The pipeline manipulations are now all relative to some known
fixed handler, e.g. user handlers are installed after the previous
user handler, the message gate is installed prior to the error handler
tail, etc.