Enforce a ratio of short-side to long-side search envelope max 1:100
To avoid bad performance when having very thin rectangles as spacial
index range queries, we ensure that the rectangle has at least a
ratio of 1:100 for the shorter and the longer sides of the rectangle/
cube. This will get more false positives that are handled during
post-filtering, but avoids having too many small search ranges for
the native index.