NPE in discovery service
The old logic was a bit odd and relied upon SocketAddress throwing an
IllegalArgumentException during parsing of null, and that was also
the only cases it could handle. This can be refactored into something
much better in the future. This is just a simple quick-fix to make
sure that all attributes are available before populating the
topology map with a CoreServerInfo.