Restore ability of BatchingNeoStores to create new store based on already existent
Because of incorrect place where transaction metadata was set previously
it was impossible to create new store with non empty initial ids
since that was blocked by count store rotation process that will wait
for a transactions that will never come since we only create empty store
with some metadata.
This PR restore that ability and adds a test that confirms that's it possible.