Spatial and Temporal don't assume file existence
Removing assumption in Spatial and Temporal index that they will at least
have some file during startup.

When starting a 3.3 store with neo4j-3.4 (with or without doing migration)
all legacy schema indexes would report initial state to be POPULATING
because Spatial and Temporal indexes assumed their files to exist on
startup which is not the case when opening store on 3.4 for the first time.
This caused all indexes to be rebuilt exactly as they where even if they
did not need recovery.

The file existence assumpion in Spatial and Temporal was meant to guard for
if files where deleted but as they both create their subindex files lazily
anyway, this guard did not cover many cases.

Also improve index migration test to verify that no recovery is done for
the old 3.3 indexes during startup.